Cleanup, improvements for statuseffects

+++ Assets/Scripts/ScriptableObjects/WaveSettings.cs
@@ -39,6 +39,6 @@ public class EnemyGroup
     [HideInInspector] public string SlotName;
     public BasicEnemy Enemy;
     public int EnemyAmount;
-    public int EnemyDelay;
-    public int GroupDelay;
+    public float EnemyDelay;
+    public float GroupDelay;
 }
